+++ internal/store/storetests/mock.go
@@ -83,7 +83,7 @@ func (m *MockStore) Get(key string, opts store.GetOptions) ([]byte, store.Stat, 
 	}, nil
 }
 
-func (m *MockStore) GetPartial(key string, start, length int, opts store.GetOptions) ([]byte, store.Stat, error) {
+func (m *MockStore) GetPartial(key string, start, length int64, opts store.GetOptions) ([]byte, store.Stat, error) {
 	d, st, err := m.Get(key, opts)
 	if err != nil {
 		return nil, store.Stat{}, err
@@ -91,14 +91,14 @@ func (m *MockStore) GetPartial(key string, start, length int, opts store.GetOpti
 	if start < 0 {
 		start = 0
 	}
-	if length < 0 || start+length > len(d) {
-		length = len(d) - start
+	if length < 0 || start+length > int64(len(d)) {
+		length = int64(len(d)) - start
 	}
 	if length <= 0 {
 		return []byte{}, st, nil
 	}
-	d2 := make([]byte, length)
-	if copy(d2, d[start:]) != length {
+	d2 := make([]byte, int(length))
+	if copy(d2, d[int(start):]) != int(length) {
 		panic("never happens")
 	}
 	return d2, st, nil
